modelo

+++ Asiservy.Automatizacion.Formularios/Asiservy.Automatizacion.Datos/Datos/MONITOREO_DESCONGELADO.cs
@@ -23,9 +23,9 @@ namespace Asiservy.Automatizacion.Datos.Datos
         public System.DateTime Hora { get; set; }
         public string Tipo { get; set; }
         public Nullable<decimal> TemperaturaAgua { get; set; }
-        public int Muestra1 { get; set; }
-        public int Muestra2 { get; set; }
-        public int Muestra3 { get; set; }
+        public decimal Muestra1 { get; set; }
+        public decimal Muestra2 { get; set; }
+        public decimal Muestra3 { get; set; }
         public string EstadoRegistro { get; set; }
         public System.DateTime FechaIngresoLog { get; set; }
         public string UsuarioIngresoLog { get; set; }
